requirements.txt
pyyaml
python-box
numpy
pandas
matplotlib
seaborn
scikit-learn
wordcloud
spacy==3.7.2
pt_core_news_md @ https://github.com/explosion/spacy-models/releases/download/pt_core_news_md-3.7.0/pt_core_news_md-3.7.0-py3-none-any.whl
streamlit
-e .